feat(api): 관심사에 따라 파일로 분리
