+++ src/api/index.ts
@@ -1,13 +1,15 @@
-import type {
-  ServerNameType,
-  ProductType,
-  CartType,
-  CouponType,
-  OrderType,
-  OrderDetailType,
-} from '../types';
+import type { ServerNameType, ProductType, CouponType } from '../types';
 
+import { postJoin, postLogin } from './auth';
+import {
+  deleteCartItem,
+  deleteCartItems,
+  getCart,
+  patchCartItemQuantity,
+  postCartItem,
+} from './cart';
 import { BASE_URL_MAP } from '../constants';
+import { getOrder, getOrders, postOrder } from './order';
 
 export const getProducts = async (serverName: ServerNameType): Promise<ProductType[]> => {
   const url = `${BASE_URL_MAP[serverName]}/products`;
@@ -17,44 +19,6 @@ export const getProducts = async (serverName: ServerNameType): Promise<ProductTy
 
   if (!response.ok) throw new Error(`${url} GET error`);
   return response.json();
-};
-
-export const postLogin = async (serverName: ServerNameType, name: string, password: string) => {
-  const url = `${BASE_URL_MAP[serverName]}/users/login`;
-  const body = JSON.stringify({
-    name,
-    password,
-  });
-
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body,
-  });
-
-  if (!response.ok) throw new Error(`${url} POST Error`);
-
-  const { token }: { token: string } = await response.json();
-
-  return token;
-};
-
-export const postJoin = (serverName: ServerNameType, name: string, password: string) => {
-  const url = `${BASE_URL_MAP[serverName]}/users/join`;
-  const body = JSON.stringify({
-    name,
-    password,
-  });
-
-  return fetch(url, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body,
-  });
 };
 
 export const getCoupons = async (
@@ -73,158 +37,19 @@ export const getCoupons = async (
   return response.json();
 };
 
-export const getCart = async (serverName: ServerNameType, token: string) => {
-  const url = `${BASE_URL_MAP[serverName]}/cart-items`;
-  const response = await fetch(url, {
-    method: 'GET',
-    headers: {
-      Authorization: `Basic ${token}`,
-    },
-  });
-
-  if (!response.ok) throw new Error(`${url} GET error`);
-
-  const cart: CartType = await response.json();
-
-  return cart;
+const api = {
+  postJoin,
+  postLogin,
+  getProducts,
+  getCoupons,
+  getCart,
+  postCartItem,
+  patchCartItemQuantity,
+  deleteCartItem,
+  deleteCartItems,
+  getOrders,
+  getOrder,
+  postOrder,
 };
 
-export const postCartItem = async (
-  serverName: ServerNameType,
-  token: string,
-  productId: number
-) => {
-  const url = `${BASE_URL_MAP[serverName]}/cart-items`;
-  const body = JSON.stringify({
-    productId,
-  });
-
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {
-      Authorization: `Basic ${token}`,
-      'Content-Type': 'application/json',
-    },
-    body,
-  });
-
-  if (!response.ok) throw new Error(`${url} POST Error`);
-};
-
-export const patchCartItemQuantity = async (
-  serverName: ServerNameType,
-  token: string,
-  cartItemId: number,
-  quantity: number
-) => {
-  const url = `${BASE_URL_MAP[serverName]}/cart-items/${cartItemId}`;
-  const body = JSON.stringify({
-    quantity,
-  });
-
-  const response = await fetch(url, {
-    method: 'PATCH',
-    headers: {
-      Authorization: `Basic ${token}`,
-      'Content-Type': 'application/json',
-    },
-    body,
-  });
-
-  if (!response.ok) throw new Error(`${url} PATCH Error`);
-};
-
-export const deleteCartItem = async (
-  serverName: ServerNameType,
-  token: string,
-  cartItemId: number
-) => {
-  const url = `${BASE_URL_MAP[serverName]}/cart-items/${cartItemId}`;
-  const response = await fetch(url, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Basic ${token}`,
-    },
-  });
-
-  if (!response.ok) throw new Error(`${url} FETCH Error`);
-};
-
-export const deleteCartItems = async (
-  serverName: ServerNameType,
-  token: string,
-  cartItemIdList: number[]
-) => {
-  const ids = cartItemIdList.map(String).join(',');
-  const url = `${BASE_URL_MAP[serverName]}/cart-items?ids=${ids}`;
-  const response = await fetch(url, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Basic ${token}`,
-    },
-  });
-
-  if (!response.ok) throw new Error(`${url} FETCH Error`);
-};
-
-export const getOrders = async (
-  serverName: ServerNameType,
-  token: string
-): Promise<OrderType[]> => {
-  const url = `${BASE_URL_MAP[serverName]}/orders`;
-  const response = await fetch(url, {
-    method: 'GET',
-    headers: {
-      Authorization: `Basic ${token}`,
-    },
-  });
-
-  if (!response.ok) throw new Error(`${url} GET error`);
-  return response.json();
-};
-
-export const getOrder = async (
-  serverName: ServerNameType,
-  token: string,
-  orderId: number
-): Promise<OrderDetailType> => {
-  const url = `${BASE_URL_MAP[serverName]}/orders/${orderId}`;
-  const response = await fetch(url, {
-    method: 'GET',
-    headers: {
-      Authorization: `Basic ${token}`,
-    },
-  });
-
-  if (!response.ok) throw new Error(`${url} GET error`);
-  return response.json();
-};
-
-interface OrderRequestItemType {
-  productId: number;
-  quantity: number;
-}
-
-export const postOrder = async (
-  serverName: ServerNameType,
-  token: string,
-  items: OrderRequestItemType[],
-  couponId: number | null
-) => {
-  const url = `${BASE_URL_MAP[serverName]}/orders`;
-  const body = JSON.stringify({
-    items,
-    couponId,
-  });
-
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {
-      Authorization: `Basic ${token}`,
-      'Content-Type': 'application/json',
-    },
-    body,
-  });
-
-  if (!response.ok) throw new Error(`${url} POST error`);
-};
+export default api;
